added test case for inline content download forms
